Bump next js version
