+++ app/(routes)/profile/page.tsx
@@ -1,10 +1,10 @@
 import { getServerSession } from "next-auth/next";
-import { authOptions } from "../api/auth/[...nextauth]/route";
 import Image from "next/image";
 
 import TopArtists from "../../_components/top_artists/TopArtists";
 import SpotifyWebApi from "spotify-web-api-node";
 import Heading from "@/app/_components/heading/Heading";
+import { authOptions } from "@/app/_utlis/auth-js-config";
 
 export default async function Profile() {
 
@@ -14,7 +14,7 @@ export default async function Profile() {
 
 
 
-  if(accessToken){
+  if (accessToken) {
     api.setAccessToken(accessToken);
 
     return (
@@ -27,16 +27,16 @@ export default async function Profile() {
         </div>
         <Image className="absolute top-24 right-16 w-96 h-80 hidden xl:inline-block" width={1000} height={1000} src={"/img/spaceman.png"} alt="spaceman"></Image>
 
-        <Heading currentView="Galaxy" alternateViews={["My Constellation", "My Globe"]} viewURLs={["/constellation", "/"]} accessToken={accessToken}/>
+        <Heading currentView="Galaxy" alternateViews={["My Constellation", "My Globe"]} viewURLs={["/constellation", "/"]} accessToken={accessToken} />
 
-        <TopArtists accessToken = {accessToken} ></TopArtists>
-        
-          
+        <TopArtists accessToken={accessToken} ></TopArtists>
+
+
       </main>
     );
   }
-  return(
+  return (
     <p>There was an error</p>
   );
-    
+
 }
