+++ package.json
@@ -14,17 +14,16 @@
     "@types/react-dom": "18.2.6",
     "d3": "^7.8.5",
     "eslint": "8.43.0",
-    "eslint-config-next": "13.4.7",
-    "next": "^13.5.6",
-    "next-auth": "^4.22.1",
+    "eslint-config-next": "^15.2.4",
+    "next": "^15.2.4",
     "rc-slider": "^10.5.0",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
     "react-range": "^1.8.14",
-    "react-slider": "^2.0.6",
     "sharp": "^0.33.0",
     "spotify-web-api-node": "^5.0.2",
-    "typescript": "5.1.3"
+    "typescript": "5.1.3",
+    "next-auth": "^4.22.1"
   },
   "devDependencies": {
     "@stylistic/eslint-plugin-ts": "^1.5.1",
